Updated guesslang and discordpy
Using guesslang again.
Guesslang is updated to 2.2.1, including tensorflow.
Discord py is updated to 1.7.3 , it is no longer locket as this is the last
version that will be supported (project is halted).

Note for server admin: Did a clean installation on server and used --no-cache-dir to fix the weird installation error
(bot for guesslang and for tensorflow)